+++ resources/views/bootstrap-4/components/forms/label.blade.php
@@ -1,6 +1,6 @@
 <label for="{{ $for }}" {{ $attributes }}>
     @if ($slot->isEmpty())
-        {{ $fallback }}
+        {{ $fallback() }}
     @else
         {{ $slot }}
     @endif