+++ resources/views/bootstrap-5/components/buttons/logout.blade.php
@@ -1,5 +1,0 @@
-{{--
-    This view serves no more purpose than that of its parent component.
-    It is only there so that it can be published and overloaded.
---}}
-@include('blade-ui-kit-bootstrap::bootstrap-5.components.buttons.form-button')